Removes phone-to-phone setup for multidevice tests
Removes the two phone setup for NFC multidevice tests, which will no
longer be supported in Android 16. A phone-to-PN532 setup is required to
run the tests.

Additional documentation cleanup as well.

This test needs to be submitted in internal Gerrit first due to an
auto-merge conflict upstream.

Bug: 395949765
Flag: EXEMPT bugfix
Test: atest CtsNfcHceMultiDeviceTestCases
Change-Id: I7907d9402bc9916c885d5751ee4f51fc97b9f799