Add NfcEmulatorApduApp to CtsNfcHceMultiDeviceTestCases.
This allows the app to be easily referenced and appended to Mobly tests in google3, as well as allow the app to be used in future multi-device tests in this suite.

This CL is a duplicate of aosp/3490032, and will be submitted internally to resolve a merge conflict.

Bug: b/395679040
Test: atest CtsNfcHceMultiDeviceTestCases
Flag: TEST_ONLY
Change-Id: I702ee0df5a6fdd8d78867f53523627c1dbd15dd0