Bypass Nfc HCE test for devices without hce feature
The bypass logic was implemented in a different way in the original
code. The method, however, will cause the test being marked as fail
during the test. According to b/388620356#comment2, this change use the
AndroidDeviceFeatureCheckDecorator in AndroidTest.xml to bypass this
module as a whole.

Bug: 399541997
Test: build and run cts-v-host
Flag: EXEMPT test fix
Change-Id: I0412d44fd3e01fc998823728f8f6b3f3f5f91c7e